develop/modules/sender-module/src/requirements.txt
confluent-kafka[avro]
psycopg2-binary
pyyaml